シリアライザーを整理、ファイル分け
