+++ myapp/serializers/api_log_serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import ApiLog
+
+# ログ
+class ApiLogSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = ApiLog
+        fields = '__all__'
